--- Assets/Scripts/CutMove.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 337e4f7edfbd340e485fd8026fea4224
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
