+++ Assets/Scripts/OperationSettings.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: ff2957cb8bfda40bbb9b746ac9c850a0
+guid: 96ec56fea9fbc4b4685d6bd931c4286f
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
